+++ CHANGELOG.md
@@ -12,9 +12,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 -->
 
 ## [Unreleased]
+### Added
+- Explicitly include `security/openpam.h` on FreeBSD
+
 ### Changed
-- Support FreeBSD by manually setting `PAM_SILENT` and conditional compilation in `wrapper.h`
 - Slightly optimize `include` setting in `Cargo.toml`
+- Split up platform-specific adaptions in `build.rs` to handle FreeBSD/NetBSD with OpenPAM
+- Bump bindgen to 0.59 (and switch from deprecated functions)
 
 ## [1.0.0-alpha3] - 2021-03-28
 ### Changed